backend/requirements.txt
numpy>=1.26.3,<2.0
pandas
pymatchmaker
fastapi
uvicorn
websockets
python-multipart
